loader.py: fix thread_context for stacks in high end
This fixes 'osv info threads' command.

when $rsp holds 0xffffc0003fdf6fc8, ulong() for gdb.Value representing
it returns 18446673706036981704. This is out of int range. This
causes that thread_context.__exit__() throws "Numeric constant too
large" exception from

  set $rbp = 18446673706036981704

The fix removes the additional conversion as it seems to be
unnecessary. The gdb.Value should format itself properly.

+++ scripts/loader.py
@@ -311,9 +311,9 @@ class thread_context(object):
         self.old_frame = gdb.selected_frame()
         self.new_frame = gdb.newest_frame()
         self.new_frame.select()
-        self.old_rsp = ulong(gdb.parse_and_eval('$rsp').cast(ulong_type))
-        self.old_rip = ulong(gdb.parse_and_eval('$rip').cast(ulong_type))
-        self.old_rbp = ulong(gdb.parse_and_eval('$rbp').cast(ulong_type))
+        self.old_rsp = gdb.parse_and_eval('$rsp').cast(ulong_type)
+        self.old_rip = gdb.parse_and_eval('$rip').cast(ulong_type)
+        self.old_rbp = gdb.parse_and_eval('$rbp').cast(ulong_type)
         self.running_cpu = state.cpu_from_thread(thread)
         self.vm_thread = gdb.selected_thread()
         if not self.running_cpu:
